+++ benchmarks/index.benchmark.ts
@@ -1,3 +1,5 @@
+require('./traversal-shallow-benchmark');
 require('./dereference-address-schema.benchmark');
 require('./dereference-petstore-swagger.benchmark');
 require('./dereference-temando-swagger.benchmark');
+
