+++ src/index.ts
@@ -2,6 +2,6 @@ import dereference from './dereference';
 import get from './get';
 
 export {
-    dereference,
-    get,
+  dereference,
+  get,
 };
